Improve names and test env. structure
