--- server/src/test/java/io/spine/server/projection/given/ProjectionTransactionTestEnv.java
@@ -1,86 +1,0 @@
-/*
- * Copyright 2019, TeamDev. All rights reserved.
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.server.projection.given;
-
-import com.google.common.collect.ImmutableList;
-import com.google.protobuf.Message;
-import io.spine.core.Subscribe;
-import io.spine.server.projection.Projection;
-import io.spine.test.projection.Project;
-import io.spine.test.projection.ProjectId;
-import io.spine.test.projection.event.PrjProjectCreated;
-import io.spine.test.projection.event.PrjTaskAdded;
-import io.spine.validate.ConstraintViolation;
-import org.checkerframework.checker.nullness.qual.Nullable;
-
-import java.util.List;
-
-import static com.google.common.collect.Lists.newLinkedList;
-
-public class ProjectionTransactionTestEnv {
-
-    /** Prevents instantiation of this utility class. */
-    private ProjectionTransactionTestEnv() {
-    }
-
-    public static class TestProjection
-            extends Projection<ProjectId, Project, Project.Builder> {
-
-        private final List<Message> receivedEvents = newLinkedList();
-        private final List<ConstraintViolation> violations;
-
-        public TestProjection(ProjectId id) {
-            this(id, null);
-        }
-
-        public TestProjection(ProjectId id,
-                              @Nullable ImmutableList<ConstraintViolation> violations) {
-            super(id);
-            this.violations = violations;
-        }
-
-        @Override
-        protected List<ConstraintViolation> checkEntityState(Project newState) {
-            if (violations != null) {
-                return ImmutableList.copyOf(violations);
-            }
-            return super.checkEntityState(newState);
-        }
-
-        @Subscribe
-        void event(PrjProjectCreated event) {
-            receivedEvents.add(event);
-            Project newState = Project.newBuilder(state())
-                                      .setId(event.getProjectId())
-                                      .build();
-            builder().mergeFrom(newState);
-        }
-
-        @Subscribe
-        void event(PrjTaskAdded event) {
-            throw new RuntimeException("that tests the projection tx behaviour");
-        }
-
-        public List<Message> getReceivedEvents() {
-            return ImmutableList.copyOf(receivedEvents);
-        }
-    }
-}
